requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
tensorflow>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Framework
streamlit>=1.20.0
fastapi>=0.75.0
uvicorn>=0.17.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Model Persistence
joblib>=1.1.0

# Gradient Boosting
xgboost>=1.6.0
lightgbm>=3.3.0

# Data Processing
imbalanced-learn>=0.8.0

# Image Processing (PIL only - no OpenCV)
pillow>=8.3.0